requirements.txt
Django>=5.0,<6.0
Pillow>=9.0
pymysql>=1.0
django-cities-light